+++ packages/syntax-highlighter/src/ansi/themes/solarized.ts
@@ -1,5 +1,0 @@
-export {
-  solarizedDark as solarized,
-  solarizedDark as theme,
-  solarizedDark as default,
-} from "./solarized-dark.ts";