+++ packages/syntax-highlighter/src/core/renderer.ts
@@ -1,1 +1,0 @@
-export { CSSHighlightRenderer, HIGHLIGHT_PREFIX } from "./css-renderer.ts";